requirements.deploy.txt
Flask
Flask-Bootstrap==3.3.7.1
python-dotenv
pytz
gunicorn
